+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * **refer** Removed IDiscoverable and ILocateable
 * **refer** Changed IReferenceable interface
 * **refer** ReferenceSet was renamed to References
+* **log** Convertion methods moved from LogLevel to LogLevelConverter
 
 ## <a name="1.0.0"></a> 1.0.0 (2016-09-22)